refactor: 교차 검증 및 권한 검증 추가
- update 메서드 article 교차 검증
- delete 메서드 권한 검증 및 article 교차 검증

+++ src/main/java/com/mobble/mobbleserver/domain/comment/service/CommentService.java
@@ -2,6 +2,8 @@ package com.mobble.mobbleserver.domain.comment.service;
 
 import com.mobble.mobbleserver.domain.article.entity.Article;
 import com.mobble.mobbleserver.domain.article.validator.ArticleValidator;
+import com.mobble.mobbleserver.domain.club.core.entity.Club;
+import com.mobble.mobbleserver.domain.club.core.policy.ClubPermissionPolicy;
 import com.mobble.mobbleserver.domain.clubMember.entity.ClubMember;
 import com.mobble.mobbleserver.domain.clubMember.validator.ClubMemberValidator;
 import com.mobble.mobbleserver.domain.comment.dto.request.CommentRequestDto;
@@ -62,16 +64,41 @@ public class CommentService {
     }
 
     @Transactional
-    public CommentResponseDto updateComment(Long commentId, Long memberId, CommentRequestDto dto) {
-        Comment comment = commentValidator.findCommentByCommentIdAndMemberIdOrThrow(commentId, memberId);
+    public CommentResponseDto updateComment(
+            Long articleId,
+            Long commentId,
+            Long memberId,
+            CommentRequestDto dto
+    ) {
+        Article article = articleValidator.findArticleByArticleIdOrThrow(articleId);
+        Club club = article.getClub();
+        ClubMember clubMember = clubMemberValidator.findClubMemberByClubIdAndMemberIdOrThrow(club.getId(), memberId);
+        Member member = clubMember.getMember();
+        Comment comment = commentValidator.findCommentByCommentIdAndMemberIdOrThrow(commentId, member.getId());
+
+        if (!comment.getArticle().getId().equals(articleId)) throw new IllegalArgumentException("");
+
         Comment updatedComment = comment.updateContent(dto.content());
 
         return CommentResponseDto.toDto(updatedComment);
     }
 
     @Transactional
-    public void deleteComment(Long commentId, Long memberId) {
-        Comment comment = commentValidator.findCommentByCommentIdAndMemberIdOrThrow(commentId, memberId);
+    public void deleteComment(Long articleId, Long commentId, Long memberId) {
+        Article article = articleValidator.findArticleByArticleIdOrThrow(articleId);
+        Club club = article.getClub();
+        ClubMember clubMember = clubMemberValidator.findClubMemberByClubIdAndMemberIdOrThrow(club.getId(), memberId);
+
+        Comment comment;
+
+        if (ClubPermissionPolicy.isLeaderOrManager(clubMember)) {
+            comment = commentValidator.findCommentByCommentIdOrThrow(commentId);
+        } else {
+            comment = commentValidator.findCommentByCommentIdAndMemberIdOrThrow(commentId, memberId);
+        }
+
+        if (!comment.getArticle().getId().equals(articleId)) throw new IllegalArgumentException("");
+
         commentRepository.delete(comment);
     }
